Added auth logic on backend

+++ backend/controllers/habitController.js
@@ -1,16 +1,23 @@
 const pool = require("../db");
 
 const getHabits = async (req, res) => {
-  const result = await pool.query("SELECT * FROM habits ORDER BY id ASC");
+  const userId = req.user.userId;
+
+  const result = await pool.query(
+    "SELECT * FROM habits WHERE user_id=$1 ORDER BY id ASC",
+    [userId],
+  );
+
   res.json(result.rows);
 };
 
 const createHabit = async (req, res) => {
   const { name, type, unit } = req.body;
+  const userId = req.user.userId;
 
   const result = await pool.query(
-    "INSERT INTO habits (name, type, unit) VALUES ($1, $2, $3) RETURNING *",
-    [name, type, unit],
+    "INSERT INTO habits (name, type, unit, user_id) VALUES ($1, $2, $3, $4) RETURNING *",
+    [name, type, unit, userId],
   );
 
   res.status(201).json(result.rows[0]);
@@ -18,10 +25,11 @@ const createHabit = async (req, res) => {
 
 const deleteHabit = async (req, res) => {
   const { habitId } = req.params;
+  const userId = req.user.userId;
 
   const result = await pool.query(
-    "DELETE FROM habits WHERE id = $1 RETURNING *",
-    [habitId],
+    "DELETE FROM habits WHERE id = $1 and user_id= $2 RETURNING *",
+    [habitId, userId],
   );
 
   res.json(result.rows[0]);
